Refactor code to work with merge.
